fix fp16's autocasting type mismatch error in CheckpointFunction's backward pass

+++ ldm/modules/diffusionmodules/util.py
@@ -134,7 +134,7 @@ class CheckpointFunction(torch.autograd.Function):
             # Fixes a bug where the first op in run_function modifies the
             # Tensor storage in place, which is not allowed for detach()'d
             # Tensors.
-            shallow_copies = [x.view_as(x) for x in ctx.input_tensors]
+            shallow_copies = [x.view_as(x).float() for x in ctx.input_tensors]
             output_tensors = ctx.run_function(*shallow_copies)
         input_grads = torch.autograd.grad(
             output_tensors,
